+++ packages/core/.storybook/mocks/index.ts
@@ -1,2 +1,4 @@
 export { product } from './product'
 export { productGridItems } from './productGridItems'
+export { searchTerms } from './searchTerms'
+export { searchHistory } from './searchHistory'