+++ packages/core/.storybook/mocks/searchHistory.ts
@@ -1,0 +1,6 @@
+export const searchHistory = [
+  { term: 'headphone', path: '/' },
+  { term: 'audio & video', path: '/' },
+  { term: 'mh-7000', path: '/' },
+  { term: 'jbl go', path: '/' },
+]