+++ packages/core/src/components/search/SearchHistory/SearchHistory.tsx
@@ -7,30 +7,38 @@ import useSearchHistory from 'src/sdk/search/useSearchHistory'
 import useSearchInput from 'src/sdk/search/useSearchInput'
 import type { History } from 'src/sdk/search/useSearchHistory'
 
+import styles from '../search.module.scss'
+
 export interface SearchHistoryProps {
+  /**
+   * Array with history options.
+   */
   history?: History[]
 }
 
-const SearchHistory = ({ history = [] }: SearchHistoryProps) => {
+const SearchHistory = ({ history = [], ...otherProps }: SearchHistoryProps) => {
   const { onSearchInputSelection } = useSearchInput()
   const { searchHistory, clearSearchHistory } = useSearchHistory(history)
 
-  if (!searchHistory.length) {
+  const historyMap = searchHistory.length ? searchHistory : history
+
+  if (!historyMap.length) {
     return null
   }
 
   return (
-    <section data-fs-search-suggestion-section>
-      <div data-fs-search-suggestion-header>
-        <p data-fs-search-suggestion-title>History</p>
+    <section data-fs-search-section className={styles.fsSearch} {...otherProps}>
+      <div data-fs-search-header>
+        <p data-fs-search-title>History</p>
         <Button variant="tertiary" onClick={clearSearchHistory}>
           Clear History
         </Button>
       </div>
       <UIList variant="ordered">
-        {searchHistory.map((item) => (
-          <li key={item.term} data-fs-search-suggestion-item>
+        {historyMap.map((item) => (
+          <li key={item.term} data-fs-search-item>
             <Link
+              data-fs-search-item-link
               variant="display"
               href={item.path}
               onClick={() => onSearchInputSelection?.(item.term, item.path)}
@@ -39,7 +47,7 @@ const SearchHistory = ({ history = [] }: SearchHistoryProps) => {
                 name="Clock"
                 width={18}
                 height={18}
-                data-fs-search-suggestion-item-icon
+                data-fs-search-item-icon
               />
               {item.term}
             </Link>